shared/credential-manager-client/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
sniffio==1.3.1
typing_extensions>=4.12.2
uvicorn>=0.34.0
azure.identity>=1.21.0
azure-mgmt-apimanagement>=4.0.1
fastapi